Split up tasks into separate files in new module
