+++ pyproject.toml
@@ -70,7 +70,7 @@ allow_redefinition = true
 
 [tool.pylint.messages_control]
 max-line-length = 90
-disable = []
+disable = ["duplicate-code"]
 max-args = 15
 max-branches = 15
